Struct oak_functions_enclave_service::proto::oak::attestation::v1::AmdAttestationReport

pub struct AmdAttestationReport {
    pub report_data: Vec<u8>,
    pub current_tcb: Option<TcbVersion>,
    pub reported_tcb: Option<TcbVersion>,
    pub debug: bool,
    pub initial_measurement: Vec<u8>,
    pub hardware_id: Vec<u8>,
    pub vmpl: u32,
}

Values extracted from an AMD SEV-SNP attestation report.

Fields
report_data: Vec<u8>

The custom bytes that were passed to the report when it was requested.

current_tcb: Option<TcbVersion>

Current TCB Version.

reported_tcb: Option<TcbVersion>

Reported TCB version (the version can be lower than current_tcb in case of a provisional update. See https://www.amd.com/content/dam/amd/en/documents/epyc-technical-docs/specifications/56860.pdf (Section 3.4) for more detail).

debug: bool

Whether the VM was booted in debug mode.

initial_measurement: Vec<u8>

The measurement of the initial memory and CPU state of the VM before startup. This implicitly includes the measurement of the Stage 0 firmware binary.

hardware_id: Vec<u8>

The hardware ID of the AMD SEV-SNP platform that generated the attestation report.

vmpl: u32

The VM Protection Leve (VMPL) that was active when the attestation report was generated.