Enum oak_functions_enclave_service::proto::oak::attestation::v1::text_expected_value::Type

pub enum Type {
    Skipped(VerificationSkipped),
    Regex(ExpectedRegex),
    StringLiterals(ExpectedStringLiterals),
}

Variants

Skipped(VerificationSkipped)

If the reference value was set to SkipVerification, we represent that here.

Regex(ExpectedRegex)

StringLiterals(ExpectedStringLiterals)

Implementations

impl Type

pub fn encode<B>(&self, buf: &mut B)
where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<Type>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>
where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.